+++ tests/unit/test_func3d_derivatives.py
@@ -14,7 +14,7 @@
 
 import pytest
 import numpy as np
-import pyqmc.func3d as func3d
+import pyqmc.wf.func3d as func3d
 
 
 @pytest.mark.parametrize(
